sql server support

+++ bosh-director/db/migrations/director/20120427235217_allow_multiple_releases_per_deployment.rb
@@ -18,14 +18,20 @@ Sequel.migration do
       self[:deployments_releases].insert(attrs)
     end
 
-    # Needed for mysql in order to drop column (doesn't work with other adapters)
-    # newer versions of sequel support drop_foreign_key but the version breaks tests
-    if [:mysql2, :mysql].include?(adapter_scheme)
-      run("alter table deployments drop FOREIGN KEY deployments_ibfk_1")
-    end
-
-    alter_table :deployments do
-      drop_column :release_id
+    
+    if Sequel::Model.db.database_type == :mssql
+      alter_table(:deployments) do
+        drop_foreign_key :release_id
+      end
+    else
+      # Needed for mysql in order to drop column (doesn't work with other adapters)
+      # newer versions of sequel support drop_foreign_key but the version breaks tests
+      if [:mysql2, :mysql].include?(adapter_scheme)
+        run("alter table deployments drop FOREIGN KEY deployments_ibfk_1")
+      end
+      alter_table :deployments do
+        drop_column :release_id
+      end
     end
   end
